Add files via upload
Online food delivery is to order the foods of different categories, view and add to cart food, order food . It manages all the information about Users, food categories, update food & its price, order food etc. The purpose of this project is to build an website to reduce the manual work for order the food with different categories etc. 

• This website will help in accessing the information related to the food delivery to the particular destination with great ease. The users can track the information related to their food with great ease through this website. The food information with different categories can also be obtained through this website.